+++ test/unit/vault/VaultService.test.js
@@ -74,7 +74,7 @@ function treeEntries(metaOid, extras = []) {
 function setupNoVault(ref) {
   ref.resolveRef.mockRejectedValueOnce(Object.assign(
     new Error('refs/cas/vault is not defined'),
-    { code: 'GIT_REF_NOT_FOUND' },
+    { code: ErrorCodes.GIT_REF_NOT_FOUND },
   ));
 }